+++ api/urls.py
@@ -4,6 +4,7 @@ from api.views import spark_status
 
 
 urlpatterns = patterns('',
+    url(r'^spark/', spark.list),
     url(r'^spark/status/', spark_status.checkin),
     url(r'^spark/(?P<device_id>\w+)/mode/', spark.set_mode),
     url(r'^spark/(?P<device_id>\w+)/name/', spark.set_name),